Add 'Project Ideas' page; enhance homepage content
A new `/project-ideas` page has been introduced, showcasing a collection of
project concepts for developers. This page allows for the sharing of
ideas that may not be pursued personally, enabling others to utilize
them.

To improve content discoverability, the homepage (`/`) was updated with a
prominent "Here is a good place to start" section. This new section
highlights popular existing content and provides a direct link to the
new 'Project Ideas' page. Additionally, a link to the 'My Project Ideas'
page was added to the `/lists` page for consistent navigation.